+++ x2-f2/ssm/ssm-sync-postgres/src/main/kotlin/x2/api/ssm/repo/postgres/repository/LogRepository.kt
@@ -18,6 +18,6 @@ interface LogRepository: CrudRepository<LogEntity, String> {
 }
 
 interface LogStats{
-	fun getCurrent(): Int
-	fun getTimestamp(): Int
+	fun getCurrent(): Long
+	fun getTimestamp(): Long
 }
